RM#67784 Split services
