+++ axelor-base/src/main/java/com/axelor/apps/base/service/pdf/PdfService.java
@@ -5,13 +5,4 @@ import com.axelor.meta.db.MetaFile;
 
 public interface PdfService {
   MetaFile convertImageToPdf(MetaFile metaFile) throws AxelorException;
-
-  MetaFile digitallySignPdf(
-      MetaFile metaFile,
-      MetaFile certificate,
-      String certificatePassword,
-      MetaFile imageFile,
-      String reason,
-      String location)
-      throws AxelorException;
 }
