+++ axelor-human-resource/src/main/java/com/axelor/apps/hr/db/repo/ExpenseHRRepository.java
@@ -27,6 +27,7 @@ import com.axelor.apps.hr.db.ExpenseLine;
 import com.axelor.apps.hr.exception.HumanResourceExceptionMessage;
 import com.axelor.apps.hr.service.expense.ExpenseFetchPeriodService;
 import com.axelor.apps.hr.service.expense.ExpenseLineService;
+import com.axelor.apps.hr.service.expense.ExpenseProofFileService;
 import com.axelor.apps.hr.service.expense.ExpenseToolService;
 import com.axelor.i18n.I18n;
 import com.axelor.inject.Beans;
@@ -51,7 +52,7 @@ public class ExpenseHRRepository extends ExpenseRepository {
       if (expense.getStatusSelect() == ExpenseRepository.STATUS_DRAFT) {
         Beans.get(ExpenseLineService.class).completeExpenseLines(expense);
       }
-      Beans.get(ExpenseLineService.class).convertProofFilesInPdf(expense);
+      Beans.get(ExpenseProofFileService.class).convertProofFilesInPdf(expense);
       return expense;
     } catch (Exception e) {
       TraceBackService.traceExceptionFromSaveMethod(e);
