+++ axelor-human-resource/src/main/java/com/axelor/apps/hr/module/HumanResourceModule.java
@@ -87,6 +87,8 @@ import com.axelor.apps.hr.service.expense.ExpenseLineServiceImpl;
 import com.axelor.apps.hr.service.expense.ExpenseMoveReverseServiceImpl;
 import com.axelor.apps.hr.service.expense.ExpensePaymentService;
 import com.axelor.apps.hr.service.expense.ExpensePaymentServiceImpl;
+import com.axelor.apps.hr.service.expense.ExpenseProofFileService;
+import com.axelor.apps.hr.service.expense.ExpenseProofFileServiceImpl;
 import com.axelor.apps.hr.service.expense.ExpenseRefusalService;
 import com.axelor.apps.hr.service.expense.ExpenseRefusalServiceImpl;
 import com.axelor.apps.hr.service.expense.ExpenseToolService;
@@ -192,5 +194,6 @@ public class HumanResourceModule extends AxelorModule {
     bind(ExpenseToolService.class).to(ExpenseToolServiceImpl.class);
     bind(ExpenseComputationService.class).to(ExpenseComputationServiceImpl.class);
     bind(EmployeeFileDMSService.class).to(EmployeeFileDMSServiceImpl.class);
+    bind(ExpenseProofFileService.class).to(ExpenseProofFileServiceImpl.class);
   }
 }
