+++ axelor-human-resource/src/main/resources/i18n/messages_fr.csv
@@ -810,7 +810,7 @@
 "Project planning time lines","Lignes de temps planifiés projet",,
 "Project task","Tâche",,
 "ProjectPlanningTime.Activity","Prestation",,
-"Proof file digitally signed","Fichier de justification digitalement signé",,
+"Proof file digitally signed","Fichier de justification numériquement signé",,
 "Public Holiday Planning","Planning jours fériés",,
 "Public holidays","Jours fériés",,
 "Public holidays planning","Planning des jours fériés",,
